fix uint range issue when it's used for displaying size in byte

+++ include/stacktracemodel.h
@@ -13,7 +13,7 @@ struct StackRecord {
     QString funcAddr_;
 };
 
-QString sizeToString(quint32 size);
+QString sizeToString(quint64 size);
 QString timeToString(int time);
 
 class StackTraceModel : public QAbstractTableModel {
